feat[client]: ConfirmDialog 컴포넌트 추가

+++ apps/client/src/pages/my/pets/[id]/index.tsx
@@ -5,8 +5,8 @@ import { EditPetForm } from '@/features/pet/edit/ui';
 import { Header } from '@/widgets/header';
 import { useDeletePetMutation } from '@/features/pet/delete/api/useDeletePetMutation';
 import { useRouter } from '@/app/i18n/navigation';
-import { cn } from '@/shared/lib/utils';
 import TrashIcon from '@/shared/assets/icons/TrashIcon.svg';
+import { ConfirmDialog } from '@/shared/ui';
 
 export default function MyPetPage() {
   const params = useParams();
@@ -24,16 +24,17 @@ export default function MyPetPage() {
           <Header.Title>프로필 편집</Header.Title>
         </Header.Center>
         <Header.Right>
-          <button
-            onClick={() => {
+          <ConfirmDialog
+            title={'정말 삭제 하겠습니까?'}
+            onConfirm={() => {
               deletePet(id, { onSuccess: () => router.back() });
             }}
-            className={cn(
-              'flex items-center justify-center rounded-full bg-[#333333CC] px-3.5 py-2.5'
-            )}
-          >
-            <TrashIcon />
-          </button>
+            trigger={
+              <button className="flex items-center justify-center rounded-full bg-[#333333CC] px-3.5 py-2.5">
+                <TrashIcon />
+              </button>
+            }
+          />
         </Header.Right>
       </Header>
